コントローラーの記述を簡略化。

+++ app/controllers/likes_controller.rb
@@ -1,20 +1,20 @@
 class LikesController < ApplicationController
+
   def create
-    current_user.likes.create(create_params)
     @prototype = Prototype.find(params[:prototype_id])
+    current_user.likes.create(create_params)
   end
 
   def destroy
+    @prototype = Prototype.find(params[:prototype_id])
     like = Like.find(params[:id])
-    if like.user.id == current_user.id
-      like.destroy
-      @prototype = Prototype.find(params[:prototype_id])
-    end
+    like.destroy if like.user.id == current_user.id
   end
 
   private
 
   def create_params
+    @prototype = Prototype.find(params[:prototype_id])
     params.permit(:prototype_id)
   end
